Move type check error tests in own submodule
