+++ tests/type_check/error/var_unbounded.pj
