+++ tests/type_check/error/wrong_cond_result_mismatch.pj
