+++ tests/type_check/error/wrong_type_bool_binop_mismatch.pj
